+++ Content/CR4S/_Blueprint/MonsterAI/BehaviorTree/Season/BT_HiemsFrostBoss.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:4a00fd4b3056ddb7e4b9a8fce7bd6b6d2dc73d38826eb8e8a09b5af4863c7133
-size 236781
+oid sha256:5659a107d4284563a19dd8bc1eb198ac809b4fb1f9d0c075ce25693671ed1bec
+size 241109
