+++ Content/CR4S/_Blueprint/MonsterAI/Blackboard/BB_SeasonBossMonster.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:d92cfe5bcde10f3d71ebcab8346f14c298d084e684d56501091c0a0cbee5039f
-size 7334
+oid sha256:d4b6a0c991a7fc5b8215a28028f5c3ea620062125c82096449d7fafac756e68e
+size 7383
