+++ Content/CR4S/_Blueprint/MonsterAI/BossMonster/Season/BP_CeresRainyBossMonster.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:a303de4f1d2a8bd300fa681b5c662c6dd0e19e9158fcd7c5e775fbb54ffa577b
-size 36772
+oid sha256:9201f808b7b37e32fc5f2ffe731f4c35048088a6c2b1ad82131198e31f82c939
+size 36933
