+++ Content/CR4S/_Blueprint/MonsterAI/BossMonster/Season/BP_HiemsFrostBossMonster.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:f0a94c8a9ae76e7e933f892998b4b383a314244b6701bcc0e3d8f5dcdf7d67e3
-size 39820
+oid sha256:07777e863fffe9f21fe4d162641dfda5376447fc32c299dac9acda556f49d3c2
+size 39985
