+++ Content/CR4S/_Blueprint/MonsterAI/BossMonster/Season/BP_VitaDryBossMonster.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:565ddbfa605e8b55b353ffbed080365abe909cdd7cdc73804baea35923d05eaa
-size 36969
+oid sha256:e117806640a39438b8a82b45ebebf64ccd61bcb378db8d0244034378dc5d1015
+size 38040
