+++ Content/CR4S/_Blueprint/MonsterAI/EQS/Season/EQS_MeleeTelpo.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:f8bc7b6d5a9539e215b7d1a7c6f0cbbd65902c67c4dff0c507b250c67e7debed
-size 16490
+oid sha256:2630067e67cb98dc760b27bef8a72e9423aa09735bed6c8ebe40e2cc48b98c38
+size 16499
