+++ Content/CR4S/_Blueprint/MonsterAI/Skill/BP_ColdFairy.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:3ae760398667fd3a7fa6b0908747d317c2337c0d33de40ed72b3c476bcef12af
-size 35095
+oid sha256:967d3798cd26df7609e8d2129647964c1f76dff63c8170e09bf41337abba7847
+size 35240
