+++ Content/CR4S/_Blueprint/MonsterAI/Skill/BP_MadFireBall.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:f2b6ae72df5013afccaa459888774c839e92e6eeebbc416bc40abb1a5f9fea9e
-size 36430
+oid sha256:e7513d2f2eb9d4f8e75d3fb17541686730c216ec20bf71989b972771df32a93f
+size 37021
