+++ Content/CR4S/_Blueprint/MonsterAI/Skill/BP_MudField.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:05e0619e010ace469a7a9ab2344d6369a3e536254111251e7b81c2c56368ac4c
-size 38902
+oid sha256:e8ed9a36d053bace83e231e5a292944120a39c6a8ebf0bd88c608910266ca74b
+size 40163
