+++ Content/CR4S/_Blueprint/MonsterAI/Skill/BP_ThornField.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:1acb5bec97243ed31e2c206d00ea8a702ae491e80550f692cbb621c65c30bd2b
-size 40092
+oid sha256:ff24fe098ace0a188e7e0bf81b9404fa4160b12f342d4a04f52138eddd6adf85
+size 40323
