+++ Source/CR4S/MonsterAI/Task/Season/BTTaskNode_Teleport.cpp
@@ -103,10 +103,12 @@ EBTNodeResult::Type UBTTaskNode_Teleport::ExecuteTask(
         
         if (MoveResult == EPathFollowingRequestResult::Failed)
         {
+            CleanupTeleport();
             return EBTNodeResult::Failed;
         }
         else if (MoveResult == EPathFollowingRequestResult::AlreadyAtGoal)
         {
+            CleanupTeleport();
             FinishLatentTask(OwnerComp, EBTNodeResult::Succeeded);
             return EBTNodeResult::Succeeded;
         }
@@ -116,6 +118,7 @@ EBTNodeResult::Type UBTTaskNode_Teleport::ExecuteTask(
     }
     else
     {
+        CleanupTeleport();
         return EBTNodeResult::Failed;
     }
 
@@ -126,46 +129,7 @@ EBTNodeResult::Type UBTTaskNode_Teleport::AbortTask(
     UBehaviorTreeComponent& OwnerComp,
     uint8* NodeMemory)
 {
-    if (IsValid(CachedPawn))
-    {
-        if (AAIController* AIC = Cast<AAIController>(CachedPawn->GetController()))
-        {
-            if (UPathFollowingComponent* PathFollowComp = AIC->GetPathFollowingComponent())
-                PathFollowComp->OnRequestFinished.RemoveAll(this);
-        }
-
-        if (UCharacterMovementComponent* MoveComp = Cast<UCharacterMovementComponent>(CachedPawn->GetMovementComponent()))
-        {
-            MoveComp->StopMovementImmediately();
-            MoveComp->SetMovementMode(OriginalMovementMode);
-            if (OriginalMovementMode == MOVE_Walking)
-                MoveComp->MaxWalkSpeed = OriginalMaxSpeed;
-            else
-                MoveComp->MaxFlySpeed = OriginalMaxSpeed;
-            MoveComp->MaxAcceleration = OriginalAcceleration;
-        }
-
-        if (IsValid(CachedMeshComp))
-        {
-            CachedMeshComp->SetVisibility(true, true);
-        }
-        
-        if (UCapsuleComponent* CapsuleComp = CachedPawn->FindComponentByClass<UCapsuleComponent>())
-            CapsuleComp->SetCollisionResponseToChannel(ECC_Visibility, ECR_Block);
-    }
-    
-    if (TrailEffectComp)
-    {
-        TrailEffectComp->DestroyComponent();
-        TrailEffectComp = nullptr;
-    }
-    if (BodyEffectComp)
-    {
-        BodyEffectComp->DestroyComponent();
-        BodyEffectComp = nullptr;
-    }
-    
-
+    CleanupTeleport();
     return EBTNodeResult::Aborted;
 }
 
@@ -173,32 +137,42 @@ void UBTTaskNode_Teleport::OnMoveCompleted(FAIRequestID RequestID, const FPathFo
 {
     if (!IsValid(CachedOwnerComp)) return;
     
+    CleanupTeleport();
+    FinishLatentTask(*CachedOwnerComp.Get(), EBTNodeResult::Succeeded);
+}
+
+void UBTTaskNode_Teleport::CleanupTeleport()
+{
     if (IsValid(CachedPawn))
     {
         if (AAIController* AIC = Cast<AAIController>(CachedPawn->GetController()))
         {
-            if (UPathFollowingComponent* PathFollowComp = AIC->GetPathFollowingComponent())
-                PathFollowComp->OnRequestFinished.RemoveAll(this);
+            if (UPathFollowingComponent* PathFollow = AIC->GetPathFollowingComponent())
+                PathFollow->OnRequestFinished.RemoveAll(this);
         }
-        
+    }
+
+    if (IsValid(CachedPawn))
+    {
         if (UCharacterMovementComponent* MoveComp = Cast<UCharacterMovementComponent>(CachedPawn->GetMovementComponent()))
         {
             MoveComp->StopMovementImmediately();
             MoveComp->SetMovementMode(MOVE_Walking);
-            
             if (OriginalMovementMode == MOVE_Walking)
                 MoveComp->MaxWalkSpeed = OriginalMaxSpeed;
             else
-                MoveComp->MaxFlySpeed  = OriginalMaxSpeed;
+                MoveComp->MaxFlySpeed = OriginalMaxSpeed;
             MoveComp->MaxAcceleration = OriginalAcceleration;
         }
-        
-        if (IsValid(CachedMeshComp))
-        {
-            CachedMeshComp->SetVisibility(true, true);
-        }
-        if (UCapsuleComponent* CapsuleComp = CachedPawn->FindComponentByClass<UCapsuleComponent>())
-            CapsuleComp->SetCollisionResponseToChannel(ECC_Pawn, ECR_Block);
+    }
+
+    if (IsValid(CachedMeshComp))
+        CachedMeshComp->SetVisibility(true, true);
+
+    if (IsValid(CachedPawn))
+    {
+        if (auto* Capsule = CachedPawn->FindComponentByClass<UCapsuleComponent>())
+            Capsule->SetCollisionResponseToChannel(ECC_Pawn, ECR_Block);
     }
     
     if (IsValid(TrailEffectComp))
@@ -211,6 +185,4 @@ void UBTTaskNode_Teleport::OnMoveCompleted(FAIRequestID RequestID, const FPathFo
         BodyEffectComp->DestroyComponent();
         BodyEffectComp = nullptr;
     }
-    
-    FinishLatentTask(*CachedOwnerComp.Get(), EBTNodeResult::Succeeded);
 }
